move Elm to own dir
