+++ packages/quicktype-core/src/language/Elm/constants.ts
@@ -1,0 +1,40 @@
+export const forbiddenNames = [
+    "if",
+    "then",
+    "else",
+    "case",
+    "of",
+    "let",
+    "in",
+    "infix",
+    "type",
+    "module",
+    "where",
+    "import",
+    "exposing",
+    "as",
+    "port",
+    "int",
+    "float",
+    "bool",
+    "string",
+    "Jenc",
+    "Jdec",
+    "Jpipe",
+    "always",
+    "identity",
+    "Array",
+    "List",
+    "Dict",
+    "Maybe",
+    "map",
+    "toList",
+    "makeArrayEncoder",
+    "makeDictEncoder",
+    "makeNullableEncoder",
+    "Int",
+    "True",
+    "False",
+    "String",
+    "Float"
+] as const;
