+++ packages/quicktype-core/src/language/Elm/index.ts
@@ -1,0 +1,2 @@
+export { ElmTargetLanguage, elmOptions } from "./language";
+export { ElmRenderer } from "./ElmRenderer";
